feat(db-core): 카운트 칼럼 타입 수정

+++ core/core-api/src/main/java/com/imstargg/core/api/controller/v1/response/BattleEventBrawlersResultStatisticsResponse.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public record BattleEventBrawlersResultStatisticsResponse(
         List<Long> brawlerBrawlStarsIds,
-        int victoryCount,
-        int defeatCount,
-        int drawCount
+        long victoryCount,
+        long defeatCount,
+        long drawCount
 ) {
 
     public static BattleEventBrawlersResultStatisticsResponse of(BattleEventBrawlersResultStatistics statistics) {
