+++ core/core-domain/src/main/java/com/imstargg/core/domain/statistics/BattleEventBrawlerResultStatistics.java
@@ -2,9 +2,9 @@ package com.imstargg.core.domain.statistics;
 
 public record BattleEventBrawlerResultStatistics(
         long brawlerBrawlStarsId,
-        int victoryCount,
-        int defeatCount,
-        int drawCount,
-        int starPlayerCount
+        long victoryCount,
+        long defeatCount,
+        long drawCount,
+        long starPlayerCount
 ) {
 }
